+++ src/components/Pagination.vue
@@ -78,6 +78,7 @@
         <li
           :class="{
             [suit('item')]: true,
+            [suit('item', 'page')]: true,
             [suit('item', 'selected')]: state.currentRefinement === page
           }"
           v-for="page in state.pages"
@@ -174,7 +175,14 @@ export default {
   name: 'AisPagination',
   mixins: [
     createSuitMixin({ name: 'Pagination' }),
-    createWidgetMixin({ connector: connectPagination }),
+    createWidgetMixin(
+      {
+        connector: connectPagination,
+      },
+      {
+        $$widgetType: 'ais.pagination',
+      }
+    ),
     createPanelConsumerMixin({
       mapStateToCanRefine: state => state.nbPages > 1,
     }),