+++ src/components/VoiceSearch.vue
@@ -84,13 +84,28 @@ import { createWidgetMixin } from '../mixins/widget';
 export default {
   name: 'AisVoiceSearch',
   mixins: [
-    createWidgetMixin({ connector: connectVoiceSearch }),
+    createWidgetMixin(
+      {
+        connector: connectVoiceSearch,
+      },
+      {
+        $$widgetType: 'ais.voiceSearch',
+      }
+    ),
     createSuitMixin({ name: 'VoiceSearch' }),
   ],
   props: {
     searchAsYouSpeak: {
       type: Boolean,
       required: false,
+      default: undefined,
+    },
+    language: {
+      type: String,
+      default: undefined,
+    },
+    additionalQueryParameters: {
+      type: Object,
       default: undefined,
     },
     buttonTitle: {
@@ -123,6 +138,8 @@ export default {
     widgetParams() {
       return {
         searchAsYouSpeak: this.searchAsYouSpeak,
+        language: this.language,
+        additionalQueryParameters: this.additionalQueryParameters,
       };
     },
     errorNotAllowed() {